refactor!: Move infractions to core plugin
BREAKING-CHANGE: This change moves the plugin-related infraction code to the core plugin

The reason behind this change is because of the increasing reliance on auto-moderation features in Kobalt; and while this *could* be made a part of a new, AutoMod-focused plugin, for the time being (and, probably forever), this code will reside within the core, and there are no plans to move moderation to its own plugin at this time.

YAGPDB is an example of a plugin-based bot that supports this kind of architecture, and may serve as the basis as for how to handle linking automod configurations to be automatically deleted when they're no longer needed, but perhaps a different approach could be taken.