fix: Delete purged keys
In the `PurgeByUserAsync` method of the `MessagePurgeService`, a certain change has been made for efficiency and optimization. Instead of calling `db.HashGetAllAsync` directly with the input strings, a key is first created and later used for this method. Furthermore, the variable `cyclicDeletions` has been transformed into an array for immediate use. The advantage of these changes is that MongoDB does no longer have to iterate over the entire collection of purged keys each time it is interrogated.

The function for deleting purged keys has been updated. Not only bulk deletions but also cyclic deletions are now considered when deleting keys from the database. The code also became cleaner, because instead of reversing a sequence and then turning it into an array, the operation of turning into an array is the last operation performed.

The last noticeable change in this code is that the final punctuation in `InvalidOperationError` was changed from an exclamation mark to a period. While this is a minor change, it contributes to a more consistent tone in our error messages.