feat!: Add rolemenu autocomplete provider
This commit adds the aforementioned autocomplete provider, and a new `GetGuildRoleMenusRequest` class.

BREAKING: In regards to the last commit, under most circumstances, role menus are acquired by their primary key, however in cases that this information is unavialable (such as is the case in the role menu service, see below), it's necessary to use the message ID to acquire the role menu instead.

I've re-added this functionality via a new API, opting to change the existing one to work with a numeric ID for consistency's sake.