refactor!: Merge plugins into core
This PR 86s Kobalt's initial direction of being a plugin-based bot (as far as DX goes; user plugins are still on the table). The issue is that Blazor does not support a plugin architecture in a trivial manner.

Plugin routing requires implementing our own router, and using Razor Class Libraries did not appear functional in preliminary testing (https://github.com/VelvetToroyashi/BlazorPluginTesting).

BREAKING CHANGE:

This is a massive upheaval of the current project's architecture, and should be treated as such. A PR will be opened in respondance to this commit, and external collaborators will be queried for reviews.

In addition to funneling plugins into the "core" project (previous Kobalt.Core), `Kobalt.Plugins.Core.Data` has been refactored into `Kobalt.Bot.Data`; migrations had to be dumped due to the incompatibility of the current model state and the target model state.

In addition to this, Remora was updated to its latest version, with accompanying changes made in accordance to the new update, which introduces breaking changes in that regard.

Centralized build artifact locations were disabled due to complications with EF Core migrations. Hopefully this is fixed in a future EF Core/dotnet release.