refactor(infractions)!: Completely rewrite how infractions are handled.
The Infraction API and related endpoints have undergone significant changes. The API no longer requires nor accepts a secondary API call to evaluate rules. Instead, it now returns an array appended to the JSON response for additional infractions that were evaluated. This change improves the efficiency of the infraction handling process and is reflected across all infraction-related endpoints. Additionally, infraction rules can now be updated in bulk. This introduces a new endpoint (`PUT /infractions/:guild_id/rules`) for more efficient management of infraction rules, allowing multiple rules to be updated in a single API call.

The API also no longer relies on status codes to dictate when an infraction was updated. While this was technically correct as far as far as REST conventions go (200 OK ➜ Updated, 201 Created ➜ New).

Kobalt will also be overhauled to use Remora.Rest's infrastructure for the supporting APIs in a future commit, but I didn't want to overload one commit with changes.

Additionally, Kobalt now takes into account *all* infraction rule results, instead of just the first one. This may be reverted in a future case, but also prevents the case where a lesser rule (e.g. 3 mutes ➜ mute for a week) would take precedence over a greater rule (e.g. 3 strikes ➜ kick).

This also brings up the question about how rules should be evaulated; right now, it's a bit daft to evaluate irrelevant rules (e.g. checking mutes when we're applying a strike), so this is likely considered a bug, and should be fixed in a future commit.

Some things were also fixed in this commit:
- Getting guild infraction rules now returns an array (typed as IROL<T>), which prevents multiple-enumeration on a LINQ query backed by EF Core. Oops.

BREAKING-CHANGE: The API no longer exposes an endpoint to evaluate infraction rules, as it is performed server-side before the request is fulfilled and returned.

- `InfractionRuleMatch` was removed.
- `InfractionRule`s now have user-friendly names.
- `InfractionResponsePayload` now includes the reference ID and additional infractions.
- Infraction rules now take into account all rules instead of bailing on the first match
- Infractions from infraction rule matches now come with their own reasoning, labeled with `AUTOMOD`