feat(bot): Make first-party, external services optional
This commit adds a a slew of new configuration values that can be set which in turn allow the disabling of certain portions of the bot (e.g., moderation).

I do have fears that this commit may inadvertently brick things, but the bot shouldn't *crash*. Further code review is required to see what may reference these services. (e.g. the anti raid service, which is invoked from the anti-raid responder...)