refactor: Create state AFTER checking if antiraid is enabled
This changes it so that the raid state is only initialized *if* the guild is configured to use anti-raid.

There's also now a check for if a user is already known, in which case re-calculating risk score is skipped. This could potentially be a bad thing, but that's yet to be tested.